requirements.txt
opencv-python
numpy
matplotlib
torch
torchvision
scikit-learn
sentence-transformers
faiss-cpu
transformers
